+++ README.md
@@ -39,4 +39,4 @@
 - [x] 상황에 따라 개발자가 Room DB 의존성을 주입받을지, In-Memory 의존성을 주입받을지 선택할 수 있다.
 
 ### 모듈 분리
-- [ ] 내가 만든 DI 라이브러리를 모듈로 분리한다.
+- [x] 내가 만든 DI 라이브러리를 모듈로 분리한다.
