+++ app/src/main/java/woowacourse/shopping/ViewModelAutoInjection.kt
@@ -1,18 +1,30 @@
-package woowacourse.shopping.di.component
+package woowacourse.shopping
 
+import androidx.activity.ComponentActivity
+import androidx.activity.viewModels
+import androidx.annotation.MainThread
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import woowacourse.shopping.di.annotation.DiViewModel
-import woowacourse.shopping.di.annotation.Inject
-import kotlin.reflect.KClass
-import kotlin.reflect.full.hasAnnotation
+import androidx.lifecycle.viewmodel.CreationExtras
+import com.android.di.component.DiSingletonComponent
+import com.android.di.component.DiViewModelComponent
+
+@MainThread
+inline fun <reified VM : ViewModel> ComponentActivity.viewModelsWithAutoInject(
+    noinline extrasProducer: (() -> CreationExtras)? = null,
+    noinline factoryProducer: (() -> ViewModelProvider.Factory)? = { inject() },
+): Lazy<VM> {
+    return viewModels(
+        extrasProducer = extrasProducer,
+        factoryProducer = factoryProducer,
+    )
+}
 
 fun inject(): ViewModelProvider.Factory = ViewModelComponent()
 
 private class ViewModelComponent : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        val clazz = modelClass.kotlin
-        if (!clazz.hasAnnotation<DiViewModel>()) {
+        if (!DiViewModelComponent.hasAnnotation(modelClass.kotlin)) {
             throw IllegalArgumentException(ERROR_ANNOTATION)
         }
         val constructor =
@@ -25,23 +37,12 @@ private class ViewModelComponent : ViewModelProvider.Factory {
             }.toTypedArray()
 
         val viewModel = constructor.newInstance(*parameters) as T
-        injectFields(viewModel)
+        DiViewModelComponent.injectFields(viewModel)
         return viewModel
-    }
-
-    private fun <T : Any> injectFields(instance: T) {
-        val clazz = instance::class.java
-        clazz.declaredFields.forEach { field ->
-            if(field.isAnnotationPresent(Inject::class.java)){
-                field.isAccessible = true
-                val fieldInstance = DiSingletonComponent.match(field.type.kotlin)
-                field.set(instance,fieldInstance)
-            }
-        }
     }
 
     companion object {
         private const val ERROR_CONSTRUCTOR = "No suitable constructor for %s"
-        private const val  ERROR_ANNOTATION =" ViewModel class must be annotated with @DiViewModel"
+        private const val ERROR_ANNOTATION = " ViewModel class must be annotated with @DiViewModel"
     }
 }
