+++ app/src/main/java/woowacourse/shopping/data/CartRepositoryImpl.kt
@@ -2,14 +2,12 @@ package woowacourse.shopping.data
 
 import woowacourse.shopping.data.mapper.toDomain
 import woowacourse.shopping.data.mapper.toEntity
-import woowacourse.shopping.di.annotation.InMemoryDatabase
-import woowacourse.shopping.di.annotation.Qualifier
-import woowacourse.shopping.di.annotation.RoomDatabase
+import com.android.di.annotation.RoomDatabase
 import woowacourse.shopping.domain.CartRepository
 import woowacourse.shopping.model.Product
 
 class CartRepositoryImpl(
-    @RoomDatabase
+    @com.android.di.annotation.RoomDatabase
 //    @InMemoryDatabase :: can use
     private val dao: CartProductDao,
 ) : CartRepository {
