+++ app/src/main/java/woowacourse/shopping/data/di/RepositoryModule.kt
@@ -1,8 +1,8 @@
-package woowacourse.shopping.di
+package woowacourse.shopping.data.di
 
 import woowacourse.shopping.data.CartRepositoryImpl
 import woowacourse.shopping.data.ProductRepositoryImpl
-import woowacourse.shopping.di.component.DiSingletonComponent
+import com.android.di.component.DiSingletonComponent
 import woowacourse.shopping.domain.CartRepository
 import woowacourse.shopping.domain.ProductRepository
