+++ app/src/main/java/woowacourse/shopping/ui/MainActivity.kt
@@ -7,7 +7,7 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityMainBinding
-import woowacourse.shopping.di.component.viewModelsWithAutoInject
+import woowacourse.shopping.viewModelsWithAutoInject
 import woowacourse.shopping.ui.cart.CartActivity
 
 class MainActivity : AppCompatActivity() {
