+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -5,7 +5,7 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityCartBinding
-import woowacourse.shopping.di.component.viewModelsWithAutoInject
+import woowacourse.shopping.viewModelsWithAutoInject
 
 class CartActivity : AppCompatActivity() {
     private val binding by lazy { ActivityCartBinding.inflate(layoutInflater) }
