+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -5,14 +5,14 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
-import woowacourse.shopping.di.annotation.DiViewModel
-import woowacourse.shopping.di.annotation.Inject
+import com.android.di.annotation.DiViewModel
+import com.android.di.annotation.Inject
 import woowacourse.shopping.domain.CartRepository
 import woowacourse.shopping.model.Product
 
-@DiViewModel
+@com.android.di.annotation.DiViewModel
 class CartViewModel: ViewModel() {
-    @Inject
+    @com.android.di.annotation.Inject
     private lateinit var cartRepository: CartRepository
 
     private val _cartProducts: MutableLiveData<List<Product>> =
