+++ settings.gradle.kts
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "android-di"
 include(":app")
+include(":di")
